extracts password change form to reusable component and intercepts using next 14 convention to render as a modal
